# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KKamera API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [users]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/upload-target:
    get:
      operationId: getUploadTarget
      tags: [users]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current upload-target default
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadTarget"
    put:
      operationId: setUploadTarget
      tags: [users]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadTargetInput"
      responses:
        "200":
          description: Updated upload-target default
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadTarget"
...
  /users/me/export:
    get:
      operationId: exportMyData
      tags: [users]
      summary: Export all personal data for the current user (GDPR)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Personal data export
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDataExport"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    UserDataExport:
      type: object
      required: [exportedAt, user, uploadTarget, cloudConnections, referrals, referredBy, uploads, feedback]
      properties:
        exportedAt:
          type: string
        user:
          $ref: "#/components/schemas/User"
        uploadTarget:
          $ref: "#/components/schemas/UploadTarget"
        subscription:
          oneOf:
            - $ref: "#/components/schemas/ExportSubscription"
            - type: "null"
        cloudConnections:
          type: array
          description: Connection metadata only — credentials and tokens are never exported.
          items:
            $ref: "#/components/schemas/ExportCloudConnection"
        referrals:
          type: array
          items:
            $ref: "#/components/schemas/Referral"
        referredBy:
          type: array
          description: Referral rows in which this user is the referred party.
          items:
            $ref: "#/components/schemas/ExportReferredBy"
        uploads:
          type: array
          items:
            $ref: "#/components/schemas/ExportUpload"
        feedback:
          type: array
          items:
            $ref: "#/components/schemas/FeedbackItem"
    ExportSubscription:
      type: object
      required: [status, freeYearsAwarded]
      properties:
        status:
          type: string
        trialStart:
          type: ["string", "null"]
        trialEnd:
          type: ["string", "null"]
        currentPeriodEnd:
          type: ["string", "null"]
        freeYearsAwarded:
          type: integer
        createdAt:
          type: ["string", "null"]
    ExportCloudConnection:
      type: object
      required: [id, type, name, active]
      properties:
        id:
          type: integer
        type:
          type: string
        provider:
          type: ["string", "null"]
        name:
          type: string
        host:
          type: ["string", "null"]
        port:
          type: ["integer", "null"]
        username:
          type: ["string", "null"]
        uploadPath:
          type: ["string", "null"]
        accountLabel:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: ["string", "null"]
    ExportReferredBy:
      type: object
      required: [id, status]
      properties:
        id:
          type: integer
        status:
          type: string
        createdAt:
          type: ["string", "null"]
    ExportUpload:
      type: object
      required: [id, fileName, fileType, status]
      properties:
        id:
          type: integer
        fileName:
          type: string
        fileType:
          type: string
        status:
          type: string
        createdAt:
          type: ["string", "null"]
    FeedbackItem:
      type: object
      required: [id, type, message, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [bug, feature, other]
        message:
          type: string
        createdAt:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    UploadTarget:
      type: object
      required: [mode, connectionIds]
      properties:
        mode:
          type: string
          enum: [all, selected, none]
        connectionIds:
          type: array
          items:
            type: integer
    UploadTargetInput:
      type: object
      required: [mode]
      properties:
        mode:
          type: string
          enum: [all, selected, none]
        connectionIds:
          type: array
          items:
            type: integer
    User:
      type: object
      required: [id, email, name, referralCode, twoFAEnabled, onboardingCompleted, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        referralCode:
          type: string
        twoFAEnabled:
          type: boolean
        onboardingCompleted:
          type: boolean
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        onboardingCompleted:
          type: boolean
...
    Referral:
      type: object
      required: [id, referredName, status, createdAt]
      properties:
        id:
          type: integer
        referredName:
          type: string
        status:
          type: string
          enum: [pending, completed]
        createdAt:
          type: string
...