# lib/api-zod/src/generated/types/userDataExport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KKamera API
 * OpenAPI spec version: 0.1.0
 */
import type { ExportCloudConnection } from "./exportCloudConnection";
import type { ExportReferredBy } from "./exportReferredBy";
import type { ExportSubscription } from "./exportSubscription";
import type { ExportUpload } from "./exportUpload";
import type { FeedbackItem } from "./feedbackItem";
import type { Referral } from "./referral";
import type { UploadTarget } from "./uploadTarget";
import type { User } from "./user";

export interface UserDataExport {
  exportedAt: string;
  user: User;
  uploadTarget: UploadTarget;
  subscription?: ExportSubscription | null;
  /** Connection metadata only — credentials and tokens are never exported. */
  cloudConnections: ExportCloudConnection[];
  referrals: Referral[];
  /** Referral rows in which this user is the referred party. */
  referredBy: ExportReferredBy[];
  uploads: ExportUpload[];
  feedback: FeedbackItem[];
}